# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для авторизации пользователей и получения фильмов/жанров/актеров.",
        "title": "Kinopoisk API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5458",
    "basePath": "/api",
    "paths": {
        "/actors/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actors"
                ],
                "summary": "Get actor by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Actor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ActorPage"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/actors/{id}/films": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actors"
                ],
                "summary": "Get films by actor ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Actor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.MainPageFilm"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ActorPage": {
            "type": "object",
            "required": [
                "age",
                "birth_date",
                "birth_place",
                "films_number",
                "height",
                "id",
                "marital_status",
                "original_name",
                "photo",
                "russian_name",
                "zodiac_sign"
            ],
            "properties": {
                "age": {
                    "type": "integer"
                },
                "birth_date": {
                    "type": "string"
                },
                "birth_place": {
                    "type": "string"
                },
                "films_number": {
                    "type": "integer"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "marital_status": {
                    "type": "string"
                },
                "original_name": {
                    "type": "string"
                },
                "photo": {
                    "type": "string"
                },
                "russian_name": {
                    "type": "string"
                },
                "zodiac_sign": {
                    "type": "string"
                }
            }
        },
...
        "models.MainPageFilm": {
            "type": "object",
            "required": [
                "cover",
                "genre",
                "id",
                "rating",
                "title",
                "year"
            ],
            "properties": {
                "cover": {
                    "type": "string"
                },
                "genre": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
    }
}